feat: notifySubscribers prop on publish mutation
