+++ packages/subscriptions/graphql/schema-types.js
@@ -115,5 +115,14 @@ type NotificationContent {
   icon: String
 }
 
+type WebNotification {
+  title: String!
+  body: String!
+  icon: String!
+  url: String!
+  # see https://developer.mozilla.org/en-US/docs/Web/API/Notification/tag
+  tag: String!
+}
+
 
 `
