+++ servers/publikator/server.js
@@ -10,6 +10,8 @@ const { graphql: search } = require('@orbiting/backend-modules-search')
 const { graphql: auth } = require('@orbiting/backend-modules-auth')
 const { graphql: discussions } = require('@orbiting/backend-modules-discussions')
 const { graphql: embeds } = require('@orbiting/backend-modules-embeds')
+const { graphql: maillog } = require('@orbiting/backend-modules-maillog')
+const { graphql: subscriptions } = require('@orbiting/backend-modules-subscriptions')
 
 const loaderBuilders = {
   ...require('@orbiting/backend-modules-discussions/loaders'),
@@ -58,7 +60,9 @@ const run = async (workerId, config) => {
       search,
       auth,
       discussions,
-      embeds
+      embeds,
+      subscriptions,
+      maillog
     ]
   )
